changed 16 bit integers to floats when receiving data.

+++ code/IMU 9250 to serial/mpu6050_serial_to_imu/src/mpu6050_serial_to_imu_node.cpp
@@ -148,19 +148,20 @@ int main(int argc, char** argv)
                 tf::Quaternion differential_rotation;
                 differential_rotation = zero_orientation.inverse() * orientation;
                 // get accelerometer values
-                int16_t ax = ((static_cast<uint8_t>(input[data_packet_start + 6]  << 8)) | static_cast<uint8_t>(input[data_packet_start + 9]));
-                int16_t ay = ((static_cast<uint8_t>(input[data_packet_start + 7]  << 8)) | static_cast<uint8_t>(input[data_packet_start + 10]));
-                int16_t az = ((static_cast<uint8_t>(input[data_packet_start + 8]  << 8)) | static_cast<uint8_t>(input[data_packet_start + 11]));
+                double ax = ((static_cast<uint8_t>(input[data_packet_start + 6]  << 8)) | static_cast<uint8_t>(input[data_packet_start + 9]));
+                double ay = ((static_cast<uint8_t>(input[data_packet_start + 7]  << 8)) | static_cast<uint8_t>(input[data_packet_start + 10]));
+                double az = ((static_cast<uint8_t>(input[data_packet_start + 8]  << 8)) | static_cast<uint8_t>(input[data_packet_start + 11]));
                 std::cout << "AX: " << ax << " AY: " << ay << " AZ: " << az << "\n";
+                
                 // get gyro values
-                int16_t gx = ((static_cast<uint8_t>(input[data_packet_start + 12] << 8)) | static_cast<uint8_t>(input[data_packet_start + 15]));
-                int16_t gy = ((static_cast<uint8_t>(input[data_packet_start + 13] << 8)) | static_cast<uint8_t>(input[data_packet_start + 16]));
-                int16_t gz = ((static_cast<uint8_t>(input[data_packet_start + 14] << 8)) | static_cast<uint8_t>(input[data_packet_start + 17]));
+                double gx = ((static_cast<uint8_t>(input[data_packet_start + 12] << 8)) | static_cast<uint8_t>(input[data_packet_start + 15]));
+                double gy = ((static_cast<uint8_t>(input[data_packet_start + 13] << 8)) | static_cast<uint8_t>(input[data_packet_start + 16]));
+                double gz = ((static_cast<uint8_t>(input[data_packet_start + 14] << 8)) | static_cast<uint8_t>(input[data_packet_start + 17]));
 
                 // get magnetometer values
-                int16_t mx = ((static_cast<uint8_t>(input[data_packet_start + 18] << 8)) | static_cast<uint8_t>(input[data_packet_start + 21]));
-                int16_t my = ((static_cast<uint8_t>(input[data_packet_start + 19] << 8)) | static_cast<uint8_t>(input[data_packet_start + 22]));
-                int16_t mz = ((static_cast<uint8_t>(input[data_packet_start + 20] << 8)) | static_cast<uint8_t>(input[data_packet_start + 23]));
+                double mx = ((static_cast<uint8_t>(input[data_packet_start + 18] << 8)) | static_cast<uint8_t>(input[data_packet_start + 21]));
+                double my = ((static_cast<uint8_t>(input[data_packet_start + 19] << 8)) | static_cast<uint8_t>(input[data_packet_start + 22]));
+                double mz = ((static_cast<uint8_t>(input[data_packet_start + 20] << 8)) | static_cast<uint8_t>(input[data_packet_start + 23]));
 
                 // calculate rotational velocities in rad/s
                 // http://www.i2cdevlib.com/forums/topic/106-get-angular-velocity-from-mpu-6050/
